refactor: extract shared SSE query listener utilities
- Extract changeQueryIdListenState into useChangeQueryListenState hook
- Extract queryId generation into getWorkflowRunSseQueryId util
- Add proper cleanup on mutation failure in useRunWorkflowVersion
