+++ packages/twenty-front/src/modules/sse-db-event/hooks/useListenToEventsForQuery.ts
@@ -1,10 +1,9 @@
-import { requiredQueryListenersState } from '@/sse-db-event/states/requiredQueryListenersState';
-import { useCallback, useEffect } from 'react';
+import { useChangeQueryListenState } from '@/sse-db-event/hooks/useChangeQueryListenState';
+import { useEffect } from 'react';
 import {
   type MetadataGqlOperationSignature,
   type RecordGqlOperationSignature,
 } from 'twenty-shared/types';
-import { useStore } from 'jotai';
 
 export const useListenToEventsForQuery = ({
   queryId,
@@ -15,46 +14,7 @@ export const useListenToEventsForQuery = ({
     | RecordGqlOperationSignature
     | MetadataGqlOperationSignature;
 }) => {
-  const store = useStore();
-  const changeQueryIdListenState = useCallback(
-    (
-      shouldListen: boolean,
-      targetQueryId: string,
-      targetOperationSignature:
-        | RecordGqlOperationSignature
-        | MetadataGqlOperationSignature,
-    ) => {
-      const currentRequiredQueryListeners = store.get(
-        requiredQueryListenersState.atom,
-      );
-
-      const listeningForThisQueryIsActive = currentRequiredQueryListeners.some(
-        (listener) => listener.queryId === targetQueryId,
-      );
-
-      if (shouldListen === listeningForThisQueryIsActive) {
-        return;
-      }
-
-      if (shouldListen) {
-        store.set(requiredQueryListenersState.atom, [
-          ...currentRequiredQueryListeners,
-          {
-            queryId: targetQueryId,
-            operationSignature: targetOperationSignature,
-          },
-        ]);
-      } else {
-        store.set(
-          requiredQueryListenersState.atom,
-          currentRequiredQueryListeners.filter(
-            (listener) => listener.queryId !== targetQueryId,
-          ),
-        );
-      }
-    },
-    [store],
-  );
+  const { changeQueryIdListenState } = useChangeQueryListenState();
 
   useEffect(() => {
     changeQueryIdListenState(true, queryId, operationSignature);
