+++ packages/twenty-front/src/modules/workflow/utils/getWorkflowRunSseQueryId.ts
@@ -1,0 +1,2 @@
+export const getWorkflowRunSseQueryId = (workflowRunId: string) =>
+  `workflow-run-${workflowRunId}`;
