+++ packages/twenty-front/src/modules/workflow/workflow-diagram/components/WorkflowRunSSESubscribeEffect.tsx
@@ -1,12 +1,13 @@
 import { CoreObjectNameSingular } from 'twenty-shared/types';
 import { useListenToEventsForQuery } from '@/sse-db-event/hooks/useListenToEventsForQuery';
+import { getWorkflowRunSseQueryId } from '@/workflow/utils/getWorkflowRunSseQueryId';
 
 export const WorkflowRunSSESubscribeEffect = ({
   workflowRunId,
 }: {
   workflowRunId: string;
 }) => {
-  const queryId = `workflow-run-${workflowRunId}`;
+  const queryId = getWorkflowRunSseQueryId(workflowRunId);
 
   useListenToEventsForQuery({
     queryId,
